Update pipmosaic.cpp
In the slot, no 2 blinking observed as UV data overlapped of slot one. When a number of slots are 2 then mosaicWidth and mosaicHeight variable's value changed but mosaicWidthUV not updated which calculated based on mosaicWidth.
Rest of another variable should also be changed but it does not any wrong effect.