make Dump() output consistent
nested objects are treated very differently by Dump(),
and the output syntax differs a lot between objects,
so introduce a base Descriptor class and macros to avoid
repeating stuff like the field name.

Dump() now takes an optional prefix, so that nested objects
are dumped with the same syntax but indented as needed.